backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
networkx
python-multipart
yake
pydantic
orjson
httpx
python-dotenv
pytest
scikit-learn
joblib